Create a new tab for the external resources in the createFlag modal
