+++ frontend/web/components/ExternalResourcesLinkTab.tsx
@@ -1,0 +1,157 @@
+import React, { FC, useState } from 'react'
+import MyRepositoriesSelect from './MyRepositoriesSelect'
+import ExternalResourcesTable, {
+  ExternalResourcesTableType,
+} from './ExternalResourcesTable'
+import { ExternalResource } from 'common/types/responses'
+import MyIssuesSelect from './MyIssuesSelect'
+import MyPullRequestsSelect from './MyPullRequestsSelect'
+import { useCreateExternalResourceMutation } from 'common/services/useExternalResource'
+import Constants from 'common/constants'
+import Button from './base/forms/Button'
+
+type ExternalResourcesLinkTabType = {
+  githubId: string
+  organisationId: string
+  featureId: string
+  projectId: string
+}
+
+type AddExternalResourceRowType = ExternalResourcesTableType & {
+  linkedExternalResources?: ExternalResource[]
+}
+
+type GitHubStatusType = {
+  value: number
+  label: string
+}
+
+const AddExternalResourceRow: FC<AddExternalResourceRowType> = ({
+  featureId,
+  linkedExternalResources,
+  organisationId,
+  projectId,
+  repoName,
+  repoOwner,
+}) => {
+  const [externalResourceType, setExternalResourceType] = useState<string>('')
+  const [featureExternalResource, setFeatureExternalResource] =
+    useState<string>('')
+
+  const [createExternalResource] = useCreateExternalResourceMutation()
+  const githubTypes = Object.values(Constants.resourceTypes).filter(
+    (v) => v.type === 'GITHUB',
+  )
+  return (
+    <Row className='list-item'>
+      <Flex style={{ maxWidth: '170px' }}>
+        <Select
+          size='select-md'
+          placeholder={'Select Type'}
+          onChange={(v: GitHubStatusType) => setExternalResourceType(v.label)}
+          options={githubTypes.map((e) => {
+            return { label: e.label, value: e.id }
+          })}
+        />
+      </Flex>
+      <Flex className='table-column px-3'>
+        <Flex className='ml-4'>
+          {externalResourceType ==
+          Constants.resourceTypes.GITHUB_ISSUE.label ? (
+            <MyIssuesSelect
+              orgId={organisationId}
+              onChange={(v) => setFeatureExternalResource(v)}
+              repoOwner={repoOwner}
+              repoName={repoName}
+              linkedExternalResources={linkedExternalResources!}
+            />
+          ) : externalResourceType ==
+            Constants.resourceTypes.GITHUB_PR.label ? (
+            <MyPullRequestsSelect
+              orgId={organisationId}
+              onChange={(v) => setFeatureExternalResource(v)}
+              repoOwner={repoOwner}
+              repoName={repoName}
+              linkedExternalResources={linkedExternalResources!}
+            />
+          ) : (
+            <></>
+          )}
+        </Flex>
+      </Flex>
+      <div className='table-column text-center' style={{ width: '80px' }}>
+        <Button
+          className='text-right'
+          theme='primary'
+          disabled={!externalResourceType || !featureExternalResource}
+          onClick={() => {
+            createExternalResource({
+              body: {
+                feature: parseInt(featureId),
+                metadata: { status: 'open' },
+                type: externalResourceType.toUpperCase().replace(/\s+/g, '_'),
+                url: featureExternalResource,
+              },
+              feature_id: featureId,
+              project_id: projectId,
+            }).then(() => {
+              toast('External Resource Added')
+            })
+          }}
+        >
+          Save
+        </Button>
+      </div>
+    </Row>
+  )
+}
+
+const ExternalResourcesLinkTab: FC<ExternalResourcesLinkTabType> = ({
+  featureId,
+  githubId,
+  organisationId,
+  projectId,
+}) => {
+  const [repoName, setRepoName] = useState<string>('')
+  const [repoOwner, setRepoOwner] = useState<string>('')
+  const [selectedResources, setSelectedResources] =
+    useState<ExternalResource[]>()
+
+  return (
+    <>
+      <ExternalResourcesTable
+        featureId={featureId}
+        projectId={projectId}
+        organisationId={organisationId}
+        repoOwner={repoOwner}
+        repoName={repoName}
+        setSelectedResources={(r: ExternalResource[]) =>
+          setSelectedResources(r)
+        }
+      />
+      <FormGroup>
+        <MyRepositoriesSelect
+          githubId={githubId}
+          orgId={organisationId}
+          onChange={(v) => {
+            const repoData = v.split('/')
+            setRepoName(repoData[0])
+            setRepoOwner(repoData[1])
+          }}
+        />
+        {repoName && repoOwner && (
+          <AddExternalResourceRow
+            featureId={featureId}
+            projectId={projectId}
+            organisationId={organisationId}
+            repoName={repoName}
+            repoOwner={repoOwner}
+            linkedExternalResources={selectedResources}
+          />
+        )}
+      </FormGroup>
+    </>
+  )
+}
+
+export default ExternalResourcesLinkTab
