+++ frontend/web/components/ExternalResourcesTable.tsx
@@ -1,17 +1,14 @@
-import React, { FC, useEffect, useState } from 'react'
+import React, { FC, useEffect } from 'react'
 import PanelSearch from './PanelSearch'
 import Button from './base/forms/Button'
 import Icon from './Icon'
 import {
-  useCreateExternalResourceMutation,
   useGetExternalResourcesQuery,
   useDeleteExternalResourceMutation,
 } from 'common/services/useExternalResource'
 import { ExternalResource } from 'common/types/responses'
 import Constants from 'common/constants'
 import Tooltip from './Tooltip'
-import MyIssuesSelect from './MyIssuesSelect'
-import MyPullRequestsSelect from './MyPullRequestsSelect'
 
 export type ExternalResourcesTableType = {
   featureId: string
@@ -19,21 +16,13 @@ export type ExternalResourcesTableType = {
   organisationId: string
   repoName: string
   repoOwner: string
+  setSelectedResources: (r: ExternalResource[]) => void
 }
 
 type ExternalResourceRowType = {
   featureId: string
   projectId: string
   externalResource: ExternalResource
-}
-
-type AddExternalResourceRowType = ExternalResourcesTableType & {
-  linkedExternalResources?: ExternalResource[]
-}
-
-type GitHubStatusType = {
-  value: number
-  label: string
 }
 
 const ExternalResourceRow: FC<ExternalResourceRowType> = ({
@@ -99,153 +88,29 @@ const ExternalResourceRow: FC<ExternalResourceRowType> = ({
   )
 }
 
-const AddExternalResourceRow: FC<AddExternalResourceRowType> = ({
-  featureId,
-  linkedExternalResources,
-  organisationId,
-  projectId,
-  repoName,
-  repoOwner,
-}) => {
-  const [externalResourceType, setExternalResourceType] = useState<string>('')
-  const [featureExternalResource, setFeatureExternalResource] =
-    useState<string>('')
-
-  const [createExternalResource] = useCreateExternalResourceMutation()
-  const githubTypes = Object.values(Constants.resourceTypes).filter(
-    (v) => v.type === 'GITHUB',
-  )
-  return (
-    <Row className='list-item'>
-      <Flex className='table-column' style={{ maxWidth: '170px' }}>
-        <Select
-          size='select-md'
-          placeholder={'Select Type'}
-          onChange={(v: GitHubStatusType) => setExternalResourceType(v.label)}
-          options={githubTypes.map((e) => {
-            return { label: e.label, value: e.id }
-          })}
-        />
-      </Flex>
-      <Flex className='table-column px-3'>
-        <Flex className='ml-4'>
-          {externalResourceType ==
-          Constants.resourceTypes.GITHUB_ISSUE.label ? (
-            <MyIssuesSelect
-              orgId={organisationId}
-              onChange={(v) => setFeatureExternalResource(v)}
-              repoOwner={repoOwner}
-              repoName={repoName}
-              linkedExternalResources={linkedExternalResources!}
-            />
-          ) : externalResourceType ==
-            Constants.resourceTypes.GITHUB_PR.label ? (
-            <MyPullRequestsSelect
-              orgId={organisationId}
-              onChange={(v) => setFeatureExternalResource(v)}
-              repoOwner={repoOwner}
-              repoName={repoName}
-              linkedExternalResources={linkedExternalResources!}
-            />
-          ) : (
-            <></>
-          )}
-        </Flex>
-      </Flex>
-      <div className='table-column text-center' style={{ width: '80px' }}>
-        <Button
-          className='text-right'
-          theme='primary'
-          disabled={!externalResourceType || !featureExternalResource}
-          onClick={() => {
-            createExternalResource({
-              body: {
-                feature: parseInt(featureId),
-                metadata: { status: 'open' },
-                type: externalResourceType.toUpperCase().replace(/\s+/g, '_'),
-                url: featureExternalResource,
-              },
-              feature_id: featureId,
-              project_id: projectId,
-            }).then(() => {
-              toast('External Resource Added')
-              setExternalResourceType('')
-              setFeatureExternalResource('')
-            })
-          }}
-        >
-          <Icon name='plus' width={20} fill='#fff' />
-        </Button>
-      </div>
-    </Row>
-  )
-}
-
 const ExternalResourcesTable: FC<ExternalResourcesTableType> = ({
   featureId,
-  organisationId,
   projectId,
-  repoName,
-  repoOwner,
+  setSelectedResources,
 }) => {
   const { data } = useGetExternalResourcesQuery({
     feature_id: featureId,
     project_id: projectId,
   })
-  const renderRowWithPermanentRow = (v: ExternalResource, index: number) => {
-    if (index === (data?.results.length || 0) - 1) {
-      return (
-        <>
-          <ExternalResourceRow
-            key={v.id}
-            featureId={featureId}
-            projectId={projectId}
-            externalResource={v}
-          />
-          <AddExternalResourceRow
-            key='add-external-key'
-            featureId={featureId}
-            projectId={projectId}
-            organisationId={organisationId}
-            repoName={repoName}
-            repoOwner={repoOwner}
-            linkedExternalResources={data?.results}
-          />
-        </>
-      )
-    } else if (v.type !== 'permanent') {
-      return (
-        <ExternalResourceRow
-          key={v.id}
-          featureId={featureId}
-          projectId={projectId}
-          externalResource={v}
-        />
-      )
-    } else if (v.type === 'permanent') {
-      return (
-        <AddExternalResourceRow
-          key='add-external-key'
-          featureId={featureId}
-          projectId={projectId}
-          organisationId={organisationId}
-          repoName={repoName}
-          repoOwner={repoOwner}
-        />
-      )
+
+  useEffect(() => {
+    if (data?.results) {
+      setSelectedResources(data.results)
     }
-  }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [data])
 
   return (
     <>
+      <h5>GitHub Issues and Pull Requests linked</h5>
       <PanelSearch
         className='no-pad overflow-visible'
-        title='Linked Issues and Pull Requests'
-        items={
-          data?.results.length
-            ? data?.results
-            : [{ id: 'add-external-key', type: 'permanent' }]
-        }
+        items={data?.results}
         header={
           <Row className='table-header'>
             <Flex className='table-column px-3'>Type</Flex>
@@ -257,7 +122,19 @@ const ExternalResourcesTable: FC<ExternalResourcesTableType> = ({
             </div>
           </Row>
         }
-        renderRow={renderRowWithPermanentRow}
+        renderRow={(v: ExternalResource) => (
+          <ExternalResourceRow
+            key={v.id}
+            featureId={featureId}
+            projectId={projectId}
+            externalResource={v}
+          />
+        )}
+        renderNoResults={
+          <FormGroup className='text-center'>
+            You have no external resouces linked for this feature.
+          </FormGroup>
+        }
       />
     </>
   )
