+++ frontend/web/components/MyRepositoriesSelect.tsx
@@ -27,12 +27,12 @@ const MyRepositoriesSelect: FC<MyRepositoriesSelectType> = ({
   }, [data])
 
   return (
-    <>
+    <div className='mt-4'>
       {!!data?.results.length && data?.results.length !== 1 && (
         <>
           <label className='cols-sm-2 control-label'>
             {' '}
-            GitHub Repositories
+            Link new Issue / Pull Request{' '}
           </label>
           <RepositoriesSelect
             repositories={data?.results}
@@ -40,7 +40,7 @@ const MyRepositoriesSelect: FC<MyRepositoriesSelectType> = ({
           />
         </>
       )}
-    </>
+    </div>
   )
 }
